# app/src/test/resources/api-spec.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: ""
  description: Generated server url
paths:
...
  /api/workspaces/{workspaceId}/expenses/{expenseId}:
    get:
      tags:
      - expenses-api
      operationId: getExpense
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: expenseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseDto"
    put:
      tags:
      - expenses-api
      operationId: updateExpense
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: expenseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditExpenseDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseDto"
  /api/workspaces/{workspaceId}/customers/{customerId}:
    get:
      tags:
      - customers-api
      operationId: getCustomer
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: customerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDto"
    put:
      tags:
      - customers-api
      operationId: updateCustomer
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: customerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditCustomerDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDto"
...
  /api/workspaces/{workspaceId}/expenses:
    get:
      tags:
      - expenses-api
      operationId: getExpenses
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: sortBy
        in: query
        required: false
        schema:
          type: string
          enum:
          - _NOT_SUPPORTED
      - name: "freeSearchText[eq]"
        in: query
        required: false
        schema:
          type: string
      - name: pageNumber
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: sortOrder
        in: query
        required: false
        schema:
          type: string
          enum:
          - asc
          - desc
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiPageExpenseDto"
    post:
      tags:
      - expenses-api
      operationId: createExpense
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditExpenseDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseDto"
...
  /api/workspaces/{workspaceId}/customers:
    get:
      tags:
      - customers-api
      operationId: getCustomers
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: sortBy
        in: query
        required: false
        schema:
          type: string
          enum:
          - _NOT_SUPPORTED
      - name: pageNumber
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: sortOrder
        in: query
        required: false
        schema:
          type: string
          enum:
          - asc
          - desc
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiPageCustomerDto"
    post:
      tags:
      - customers-api
      operationId: createCustomer
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditCustomerDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDto"
  /api/workspaces/{workspaceId}/categories:
    get:
      tags:
      - categories-api
      operationId: getCategories
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: sortBy
        in: query
        required: false
        schema:
          type: string
          enum:
          - _NOT_SUPPORTED
      - name: pageNumber
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: sortOrder
        in: query
        required: false
        schema:
          type: string
          enum:
          - asc
          - desc
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiPageCategoryDto"
    post:
      tags:
      - categories-api
      operationId: createCategory
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryDto"
...
components:
  schemas:
...
    EditExpenseDto:
      required:
      - currency
      - datePaid
      - originalAmount
      - title
      - useDifferentExchangeRateForIncomeTaxPurposes
      type: object
      properties:
        category:
          type: integer
          format: int64
        datePaid:
          type: string
          format: date
        title:
          type: string
        currency:
          type: string
        originalAmount:
          type: integer
          format: int64
        convertedAmountInDefaultCurrency:
          type: integer
          format: int64
        useDifferentExchangeRateForIncomeTaxPurposes:
          type: boolean
        incomeTaxableAmountInDefaultCurrency:
          type: integer
          format: int64
        attachments:
          type: array
          items:
            type: integer
            format: int64
        percentOnBusiness:
          type: integer
          format: int32
        notes:
          maxLength: 1024
          minLength: 0
          type: string
        generalTax:
          type: integer
          format: int64
    ExpenseAmountsDto:
      type: object
      properties:
        originalAmountInDefaultCurrency:
          type: integer
          format: int64
        adjustedAmountInDefaultCurrency:
          type: integer
          format: int64
    ExpenseDto:
      required:
      - attachments
      - convertedAmounts
      - currency
      - datePaid
      - id
      - incomeTaxableAmounts
      - originalAmount
      - percentOnBusiness
      - status
      - timeRecorded
      - title
      - useDifferentExchangeRateForIncomeTaxPurposes
      - version
      type: object
      properties:
        category:
          type: integer
          format: int64
        title:
          type: string
        timeRecorded:
          type: string
          format: date-time
        datePaid:
          type: string
          format: date
        currency:
          type: string
        originalAmount:
          type: integer
          format: int64
        attachments:
          type: array
          items:
            type: integer
            format: int64
        percentOnBusiness:
          type: integer
          format: int32
        notes:
          type: string
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - FINALIZED
          - PENDING_CONVERSION
          - PENDING_CONVERSION_FOR_TAXATION_PURPOSES
        generalTax:
          type: integer
          format: int64
        generalTaxRateInBps:
          type: integer
          format: int32
        generalTaxAmount:
          type: integer
          format: int64
        convertedAmounts:
          $ref: "#/components/schemas/ExpenseAmountsDto"
        incomeTaxableAmounts:
          $ref: "#/components/schemas/ExpenseAmountsDto"
        useDifferentExchangeRateForIncomeTaxPurposes:
          type: boolean
    EditCustomerDto:
      required:
      - name
      type: object
      properties:
        name:
          type: string
    CustomerDto:
      required:
      - id
      - name
      - version
      type: object
      properties:
        name:
          type: string
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
...
    CreateCategoryDto:
      required:
      - expense
      - income
      - name
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        income:
          type: boolean
        expense:
          type: boolean
    CategoryDto:
      required:
      - expense
      - income
      - name
      - version
      type: object
      properties:
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        income:
          type: boolean
        expense:
          type: boolean
...
    ApiPageExpenseDto:
      required:
      - data
      - pageNumber
      - pageSize
      - totalElements
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        data:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseDto"
...
    ApiPageCustomerDto:
      required:
      - data
      - pageNumber
      - pageSize
      - totalElements
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        data:
          type: array
          items:
            $ref: "#/components/schemas/CustomerDto"
    ApiPageCategoryDto:
      required:
      - data
      - pageNumber
      - pageSize
      - totalElements
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        data:
          type: array
          items:
            $ref: "#/components/schemas/CategoryDto"
...